+++ src/Core/src/Internal/Factory.php
@@ -286,7 +286,7 @@ final class Factory implements FactoryInterface
         array $parameters = [],
         Stringable|string|null $context = null
     ): mixed {
-        $parent = $this->scope->getParent();
+        $parent = $this->scope->getParentFactory();
 
         if ($parent !== null) {
             try {
@@ -294,6 +294,7 @@ final class Factory implements FactoryInterface
                     'current scope' => $this->scope->getScopeName(),
                     'jump to parent scope' => $this->scope->getParentScope()->getScopeName(),
                 ]);
+                /** @psalm-suppress TooManyArguments */
                 return $parent->make($alias, $parameters, $context);
             } catch (BadScopeException $e) {
                 if ($this->scope->getScopeName() !== $e->getScope()) {