+++ CHANGELOG.md
@@ -5,6 +5,9 @@
 - **Medium Impact Changes**
   - [spiral/core] Interface `Spiral\Core\Container\SingletonInterface` is deprecated,
     use `Spiral\Core\Attribute\Singleton` instead. Will be removed in v4.0.
+- **Other Features**
+  - Added `Spiral\Scaffolder\Command\InfoCommand` console command for getting information about available scaffolder 
+    commands.
 
 ## 3.11.1 - 2023-12-29